(PT-260) Add error handling to gather_facts plan
This commit updates the gather facts plan and both the sys_output_to_hash and
convert_to_fact_name puppet functions to include error handling. Error
handling logic now stops the plan from failing when individual targets fail
and instead keeps track of what error occurs and outputs it at the end.

As part of this change, the result of the plan is now always a Hash with
key/value data containing target names as keys and error data as values. In
the case that dry_run is true the plan still returns a hash with what data
would have been submitted, but now that hash also includes error data for any
targets that failed.